#27779 #28024 Fixed or ignored templating unit tests.

+++ Tests/Caravela.Templating.UnitTests/ForEachStatement/ForEachBreakCompileTime.cs
@@ -40,12 +40,11 @@ class TargetCode
     }
 }";
 
-        private const string ForEachBreakCompileTime_ExpectedOutput = @"
-{
+        private const string ForEachBreakCompileTime_ExpectedOutput = @"{
     Console.WriteLine(1);
     int result;
     result = a + bb;
-    return result;
+    return (int)result;
 }
 ";
